pyproject.toml
[project]
name = "image-convert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colour-science>=0.4.6",
    "pillow>=11.3.0",
]
